+++ controllers/toplevel/robots.js
@@ -3,7 +3,7 @@ const config = require('config');
 const sitemap = require('sitemap');
 const { noCache, sMaxAge } = require('../../middleware/cached');
 const { getBaseUrl, getAbsoluteUrl } = require('../../modules/urls');
-const { getCanonicalRoutes } = require('../helpers/route-helpers');
+const { getCanonicalRoutes } = require('../../modules/route-helpers');
 
 function initRobots(router) {
     router.get('/robots.txt', noCache, (req, res) => {
