+++ modules/route-helpers.js
@@ -2,8 +2,8 @@
 
 const { compose, concat, filter, flatMap, map, omitBy, sortBy, uniqBy } = require('lodash/fp');
 
-const contentApi = require('../../services/content-api');
-const routes = require('../routes');
+const contentApi = require('../services/content-api');
+const routes = require('../controllers/routes');
 
 const sortedUniqByPath = compose(
     sortBy('path'),
